<refact> : myporfile styled component 수정
- stylecomponents 수정
