+++ src/Router/Router.jsx
@@ -7,8 +7,8 @@ import LoginEmail from '../pages/LoginEmail/LoginEmail';
 import JoinEmail from '../pages/JoinEmail/JoinEmail';
 import HomeFeed from '../pages/HomeFeed/HomeFeed';
 import SearchUser from '../pages/SearchUser/SearchUser';
-import YourProfile from '../pages/YourProfile/YourProfile';
-import MyProfile from '../pages/MyProfile/MyProfile';
+import YourProfile from '../pages/Profile/YourProfile/YourProfile';
+import MyProfile from '../pages/Profile/MyProfile/MyProfile';
 import FollowerList from '../pages/FollowerList/FollowerList';
 import FollowingList from '../pages/FollowingList/FollowingList';
 import ModifyProfile from '../pages/ModifyProfile/ModifyProfile';
