+++ src/components/MenuBar/StyledMenuBar.jsx
@@ -4,7 +4,8 @@ const MenuBarWrapper = styled.div`
   display: flex;
   justify-content: flex-end;
   width: 390px;
-  border: 0.5px solid #dbdbdb;
+  margin-bottom: 16px;
+  border-bottom: 0.5px solid #dbdbdb;
   img {
     margin: 12px 16px;
     width: 20px;
