+++ src/components/common/TabMenu/StyledTabMenu.jsx
@@ -6,8 +6,6 @@ export const MenuWrapper = styled.ul`
   bottom: 0;
   display: flex;
   justify-content: space-around;
-  position: fixed;
-  bottom: 0;
   width: 100%;
   height: 60px;
   background: #ffffff;
