+++ src/pages/Profile/MyProfile/MyProfile.jsx
@@ -1,14 +1,14 @@
-import Nav from '../../components/Nav/Nav';
-import Product from '../../components/common/Product/Product';
-import TabMenu from '../../components/common/TabMenu/TabMenu';
-import ProfileInfo from '../../components/ProfileInfo/ProfileInfo';
-import MenuBar from '../../components/MenuBar/MenuBar';
-import PostCard from '../../components/common/PostCard/PostCard';
+import Nav from '../../../components/Nav/Nav';
+import Product from '../../../components/common/Product/Product';
+import TabMenu from '../../../components/common/TabMenu/TabMenu';
+import ProfileInfo from '../../../components/ProfileInfo/ProfileInfo';
+import MenuBar from '../../../components/MenuBar/MenuBar';
+import PostCard from '../../../components/common/PostCard/PostCard';
 import * as S from '../YourProfile/StyledYourProfile';
 
 const MyProfile = () => {
   return (
-    <>
+    <S.Container>
       <Nav type='home' />
       <S.MainWrap>
         <ProfileInfo />
@@ -20,13 +20,15 @@ const MyProfile = () => {
             <Product />
           </div>
         </S.ProductsSection>
-        <MenuBar />
         <S.PostCardWrap>
+          <MenuBar />
+          <PostCard />
+          <PostCard />
           <PostCard />
         </S.PostCardWrap>
       </S.MainWrap>
       <TabMenu />
-    </>
+    </S.Container>
   );
 };
